disable total column if not in parallel (#1087)
TLGCatalog `EGT01` filed because of this.

Talked with @Melkiades and it seems that `summarize_colvars()` does not
play nicely together with `add_overall_col()` before. Hence putted this
in the if condition.

You can test it as follows:
- `pkgload::dev_example("tm_t_summary_by")`
- assure "Add All Patients column" is TRUE
- turn on "Show summarize variables in parallel"

Currently on main it's (gracefully) failing. After this fix, all
patients flag is silently ignored.