1441 Introduce transformators parameter in modules (#1313)
https://github.com/insightsengineering/teal/issues/1441

Included `transformators` parameter in modules. This is passed to
`teal::module` that does assertions and checks on this parameter.

A food for thought: default `decorators` value is `NULL` where default
`transformators` value is `list()`. Maybe we can introduce a check, that
allows `transformators` to be `NULL` on default, and which changes
`NULL` to `list()`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>